Use npm global install strategy
This fixes a bug where adaptors installed during runtime would clobber
all other adaptors. This is by design in NPM, and not something we want
here.

By switching to using `--global` and setting our own prefix we can
install additional adaptors without worrying about a specific `npm
install` considering that a new module is the _only_ module we want.

This does leave a bit of an API crack, in that we can't rely on a single
`adaptors_path`. As NPMs global style has a different directory
structure: `bin` and `lib` vs `node_modules/.bin` and `node_modules`.

We have a work around right now to take the `adaptors_path` and assume
it's the `lib` folder, and expand it out using `../bin`, not ideal but
should avoid breaking existing configurations while we find a better way
to expose the distinction between the modules and bin stubs.