Fix installation and lookup of pre-release adaptors
When a pre-release adaptor is requested, the version number was being incorrectly
read. Resulting in the module being installed incorrectly.

Subsequently when an adaptor is installed like this, the repository has
an entry that has a version of `nil`; which in-turn breaks version
resolution for requirements like "latest".

Additionally a change was made to the version filter - "latest" no
longer includes 'pre' releases.